project/backend/requirements.txt
Django>=5.0.3
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
djangorestframework-simplejwt>=5.3.1
Pillow>=10.2.0
django_cleanup>=8.0.0
django_ratelimit>=4.1.0
redis>=5.0.1
django-redis>=5.4.0